// solauto-sdk/src/generated/types/swapType.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import { Serializer, scalarEnum } from '@metaplex-foundation/umi/serializers';

export enum SwapType {
  ExactIn,
  ExactOut,
}

export type SwapTypeArgs = SwapType;

export function getSwapTypeSerializer(): Serializer<SwapTypeArgs, SwapType> {
  return scalarEnum<SwapType>(SwapType, {
    description: 'SwapType',
  }) as Serializer<SwapTypeArgs, SwapType>;
}