// solauto-sdk/src/generated/types/tokenBalanceChangeType.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import { Serializer, scalarEnum } from '@metaplex-foundation/umi/serializers';

export enum TokenBalanceChangeType {
  None,
  PreSwapDeposit,
  PostSwapDeposit,
  PostRebalanceWithdrawSupplyToken,
  PostRebalanceWithdrawDebtToken,
}

export type TokenBalanceChangeTypeArgs = TokenBalanceChangeType;

export function getTokenBalanceChangeTypeSerializer(): Serializer<
  TokenBalanceChangeTypeArgs,
  TokenBalanceChangeType
> {
  return scalarEnum<TokenBalanceChangeType>(TokenBalanceChangeType, {
    description: 'TokenBalanceChangeType',
  }) as Serializer<TokenBalanceChangeTypeArgs, TokenBalanceChangeType>;
}